refactor: shield coverage enabled
