+++ ui/pages/confirmations/components/confirm/footer/shield-footer-coverage-indicator/shield-footer-coverage-indicator.test.tsx
@@ -13,6 +13,7 @@ jest.mock(
     useEnableShieldCoverageChecks: jest.fn(() => ({
       isEnabled: true,
       isPaused: false,
+      isShowCoverageIndicator: true,
     })),
   }),
 );
