+++ ui/pages/confirmations/components/confirm/footer/shield-footer-coverage-indicator/shield-footer-coverage-indicator.tsx
@@ -1,7 +1,4 @@
-import {
-  TransactionMeta,
-  TransactionType,
-} from '@metamask/transaction-controller';
+import { TransactionMeta } from '@metamask/transaction-controller';
 import React, { useMemo } from 'react';
 import { ConfirmInfoAlertRow } from '../../../../../../components/app/confirm/info/row/alert-row/alert-row';
 import { RowAlertKey } from '../../../../../../components/app/confirm/info/row/constants';
@@ -17,30 +14,18 @@ import {
 import { useI18nContext } from '../../../../../../hooks/useI18nContext';
 import { useConfirmContext } from '../../../../context/confirm';
 import { useEnableShieldCoverageChecks } from '../../../../hooks/transactions/useEnableShieldCoverageChecks';
-import { isSignatureTransactionType } from '../../../../utils';
-import { isCorrectDeveloperTransactionType } from '../../../../../../../shared/lib/confirmation.utils';
 import useAlerts from '../../../../../../hooks/useAlerts';
 import ShieldIconAnimation from './shield-icon-animation';
 
 const ShieldFooterCoverageIndicator = () => {
   const t = useI18nContext();
   const { currentConfirmation } = useConfirmContext<TransactionMeta>();
-  const isSignature = isSignatureTransactionType(currentConfirmation);
-  const isTransactionConfirmation = isCorrectDeveloperTransactionType(
-    currentConfirmation?.type,
-  );
   const { getFieldAlerts } = useAlerts(currentConfirmation?.id ?? '');
   const fieldAlerts = getFieldAlerts(RowAlertKey.ShieldFooterCoverageIndicator);
   const selectedAlert = fieldAlerts[0];
   const selectedAlertSeverity = selectedAlert?.severity;
 
-  const { isEnabled, isPaused } = useEnableShieldCoverageChecks();
-  const isShieldApprovalTx =
-    currentConfirmation?.type === TransactionType.shieldSubscriptionApprove;
-  const isShowShieldFooterCoverageIndicator =
-    (isSignature || isTransactionConfirmation) &&
-    (isEnabled || isPaused) &&
-    !isShieldApprovalTx; // TODO: we don't show the shield footer coverage indicator for shield approval transactions atm, remove this once ruleengine update to cover shield subscription transaction
+  const { isPaused, isShowCoverageIndicator } = useEnableShieldCoverageChecks();
 
   const animationSeverity = useMemo(() => {
     if (isPaused) {
@@ -49,7 +34,7 @@ const ShieldFooterCoverageIndicator = () => {
     return selectedAlertSeverity;
   }, [isPaused, selectedAlertSeverity]);
 
-  if (!currentConfirmation || !isShowShieldFooterCoverageIndicator) {
+  if (!currentConfirmation || !isShowCoverageIndicator) {
     return null;
   }
